feat(addon-interactions): update interaction hover icon
